% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMad.R
\name{DataMad}
\alias{DataMad}
\title{Compress the data of a vector elements}
\usage{
DataMad(x)
}
\arguments{
\item{x}{vector given to compression}
}
\value{
comparison between initial data and compress data after compressions
}
\description{
This function compress a data of a given vector by using FFT and delete hight frequency.
}
\examples{
DataMad(rnorm(1024))
}
