% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevProjectSchool-package.R
\docType{package}
\name{DevProjectSchool-package}
\alias{DevProjectSchool}
\alias{DevProjectSchool-package}
\title{DevProjectSchool: Fast Fourier Transform Application}
\description{
This package help you to analyse the specter of an audio song or compress the data as you want .
}
\author{
\strong{Maintainer}: Marc Djeumou \email{marcdjeumou4@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
