% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftMad.R
\name{DftMad}
\alias{DftMad}
\title{Calculate the discret fourier transform}
\usage{
DftMad(v)
}
\arguments{
\item{v}{the vector given to have the discret fourier transform}
}
\value{
The discret fourier transform of vector v
}
\description{
This function calculate the discret fourier transform of a given vector.
}
\examples{
DftMad(sin(5*seq(0, 2*pi, length.out=20)))
}
