% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftMad.R
\name{fftMad}
\alias{fftMad}
\title{Calculate the Fast Fourier Transform}
\usage{
fftMad(x)
}
\arguments{
\item{x}{the vector given to have a fast fourier transform}
}
\value{
the fast fourier transform of vector x
}
\description{
This function calculate the discret fourier transform of a given vector.
}
\examples{
fftMad(seq(O.2*pi, length.out = 20))
}
