% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mufi.R
\name{Mufi}
\alias{Mufi}
\title{Treat a song}
\usage{
Mufi(x)
}
\arguments{
\item{X}{mean to the path of a song take on the wave extension.}
}
\value{
Return the song treat by a play test
}
\description{
This function treat a song by using Fast Fourier Transform function of this package.
}
\examples{
fft(sin(5*seq(0, 2*pi, length.out=10)))
}
