% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecMad.R
\name{SpecMad}
\alias{SpecMad}
\title{Analyse the spectre audio}
\usage{
SpecMad(x)
}
\arguments{
\item{x}{which is the path of song to analyse}
}
\value{
the specter of song given by path
}
\description{
This package analyse the specter of a son by given path
}
\examples{
SpecMad("science-and-technology-152988.wav")
}
