% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftMad.R
\name{fftMad}
\alias{fftMad}
\title{Calculate the fast fourier transform}
\usage{
fftMad(v)
}
\arguments{
\item{v}{the vector given to have the fast fourier transform}
}
\value{
The fast fourier transform of vector
}
\description{
This function calculate the fast fourier transform of a given vector.
}
\examples{
fft(sin(5*seq(0, 2*pi, length.out=10)))
}
